rename crawler interface to base_crawler, and move to /inventory
